Rename routers removing -route suffix
