+++ routes/auth-qr-code-png.js
