+++ routes/logout.js
